requirements.txt
# Data Engineering
google-cloud-storage
google-cloud-bigquery
pandas
db-dtypes
deweydatapy

# Analysis
numpy
scipy
statsmodels

# Dashboard
streamlit
plotly
altair

# Config
python-dotenv
